Re-checkout main branch
Because the `translations` branch has nothing in it except for the
README and the openverse.pot file, the post-run actions for our
`setup-node-env` action is unable to run (it won't be able to find the
files it needs). Therefore, we need to recheckout the `main` branch.
I've gone ahead and updated the initial checkout to also be explicit
about the branch it checks out so that the intention here is clear.