Enable seeking for Jamendo tracks in blink browsers; remove message
This PR enables seeking Jamendo tracks as we now use MP32s from Jamendo and an endpoint which supports the correct headers for streaming audio. This PR also removes related strings, the blink browser ua detection hook, and the custom logic to display a notice inside of the audio track for Jamendo in blink browsers.